requirements.txt
PySide6
google-generativeai
pypdf
